backend/requirements.txt
fastapi[all]

sqlalchemy

psycopg2-binary

GeoAlchemy2

alembic

python-dotenv